New tab system & new tips section!!
Added: Tips section that also explains the formula
Changed: Level-Select & Credits are no longer under a drop-down but are instead a tab each, alongside the new tips section.
Also, the selected tab gets saved between visits!

Fixed: The site calculating the score like half a second after the page loads instead of the site immediately displaying the score.
Fixed: The black gradient not being removed when the site is narrow.
Fixed: The "Calculate!" button being stuck on green when clicking on it (same I think also happened with the levels).